Scappy-Traffic-Generator
Spoofing IBM subnet, initiating traffic with a crafted DNS payload sent to multiple victims read from a list of IP's in a textfile. If the source and destinations is reverse and the text file consist of public DNS servers this would be an amplification attack, however that's not the intention of this script. 

During compile leveraging Scapy ability to add a more specfic route which will be injected in a spoofed routing table, however at the end of the script no changes were made to the intial routing table as it's copied during compile only.